safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.5_The_if_Statement/S12.5_A11-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token '}'. Expected ':' after property name.
S12.5_A11

PASS Expected parsing failure

TEST COMPLETE
